Add pi_core_register_node_pidev() function in the source.
The pi_core_register_node_pidev() function in the source file registers a
an object of type struct pi_device, which represents a client device on the
pi bus.
Further developments in the parallel_interface will use this function to
register all the child devices of a registered host device.
The function takes a device_node, that is grabbed from the device tree and uses
that node to allocate and initialize the pi-client-device.