Add the device tree overlay file for pi-host device
This commit adds device-tree-overlay source file that needs to be loaded in
order to indicate to the PRU that a pi-host device is present on the device.
The compatible string in the device tree overlay should match the compatible
string of the of_node_id structure, registered by the platform driver.

On successful addition of the device-tree-overlay the device appears as:

root@beaglebone:~# cat /sys/devices/platform/ocp/ocp\:pi0@f00/of_node/compatible
ti,pibus0