The commit add a child device node into the PI-TEST device tree overlay source
The device tree overlay PI-TEST is used to indicate that a pi-host device is
connected to the system.
Earlier support added the pi-bus host node. This commit adds child device to
this host device node. The child nodes are considered as the child of the node
superior to it. In case of a bus-specific-host device, the child device are
the client device on the bus.

The commit adds a client device, aka pidev, to the bus, with its compatible
string as 'dc782a', as that is the device board in subject. This compatible
string must match with the device ID given be the device-drivers.

Further developments will add support for device drivers.

On getting loading the device tree overlay, the device appears in the device
sysfs :
root@beaglebone:~# tree -L 2 /sys/devices/platform/ocp/ocp\:pi0@f00/pi-0/
/sys/devices/platform/ocp/ocp:pi0@f00/pi-0/
|-- of_node -> ../../../../../firmware/devicetree/base/ocp/pi0@f00
|-- pidev
|   |-- of_node -> ../../../../../../firmware/devicetree/base/ocp/pi0@f00/adc@f00
|   |-- power
|   |-- subsystem -> ../../../../../../bus/parallel_interface
|   `-- uevent
|-- power
|   |-- async
|   |-- autosuspend_delay_ms
|   |-- control
|   |-- runtime_active_kids
|   |-- runtime_active_time
|   |-- runtime_enabled
|   |-- runtime_status
|   |-- runtime_suspended_time
|   `-- runtime_usage
|-- subsystem -> ../../../../../bus/parallel_interface
`-- uevent

On successful registration of the pidev with the core bus driver, the device
appears in the bus sysfs :

root@beaglebone:~# tree -L 2 /sys/bus/parallel_interface/
/sys/bus/parallel_interface/
|-- devices
|   |-- pi-0 -> ../../../devices/platform/ocp/ocp:pi0@f00/pi-0
|   `-- pidev -> ../../../devices/platform/ocp/ocp:pi0@f00/pi-0/pidev
|-- drivers
|-- drivers_autoprobe
|-- drivers_probe
`-- uevent

pidev is the device node in subject in both the output.

On successful registration of the pidev with the bus-core driver, the device
appears as a child device in the bus sysfs entries :