Add support for pi_register_driver() and pi_unregister_driver() API
This commit adds support for:

1. The pi_register_driver() API that needs to be used by a device driver to
register itself onto the core-bus driver. This API is implemented using a few
macros to make it easy to use.

2. The pi_unregister_driver() API that needs to be used by the device driver to
unregister itself from the core-bus driver. This API is really minimal, as a
result, it has been declared inline in the parallel_interface.h header file
itself.