Add definition of the pi_device_id structure to the parallel_interface.h
This commit adds definition of the pi_device_id structure that will used by the
device-driver to register themselves to the pi-bus core.
The device drivers need to use this data structure to provide with and array of
supported hardware boards and provide with the board-specific data for that
board.

If any of the name value provided in the id table will match the compatible
string of the device, the driver will be probed.