Add support to match registered pidrv against registered pidev in core driver
Whenever a new device or device-driver is registered to a bus, the kernel uses
bus-specific match function to allow the core bus driver to implement a
comparison of the registered device against a registered device driver.

This commit implements a comparison inside the bus-specific match function to
compare if the device-driver can handle the give device. If it can be handled,
the implementation further binds the two of them.

On successful binding of the driver with the device, it can be seen in the bus
sysfs :

root@beaglebone:~# tree /sys/bus/parallel_interface/
/sys/bus/parallel_interface/
|-- devices
|   |-- pi-0 -> ../../../devices/platform/ocp/ocp:pi0@f00/pi-0
|   `-- pidev -> ../../../devices/platform/ocp/ocp:pi0@f00/pi-0/pidev
|-- drivers
|   `-- beaglescope_driver
|       |-- bind
|       |-- module -> ../../../../module/beaglescope_driver
|       |-- pidev -> ../../../../devices/platform/ocp/ocp:pi0@f00/pi-0/pidev
|       |-- uevent
|       `-- unbind
|-- drivers_autoprobe
|-- drivers_probe
`-- uevent

7 directories, 6 files

In the output, it can be seen that the beaglescope_driver is bound to the pidev
device.