Add support to probe and remove bound device-drivers
Whenever a successful match of a device driver against a device is found, the
bus specific probe method is called by the kernel. Whenever a bound device is
removed from the bus, the bus specific remove method is called.

This commit adds support to probe and remove the device drivers that are
registered to the pi bus core, and are bound to a device. This is done by
implementing the bus-specific probe and remove methods of the
parallel_interface.c source file. The method further call the device driver's
probe that is associated with this device.