Separate dc782a.c and beaglescope_driver.c file from a kind of combined source.
The BeagleScope project aims at developing generic parallel interface. As a
first step the aim was to develop a combined driver. This combined driver was
developed and named beaglescope_driver. If one would like to use just the end
product, he can use beaglescope_driver.

As the next step, the work on parallel interface has started and
beaglescope_driver with some modification, was used as the device driver.

This commit now separates out the beaglescope_driver and the actual device
driver dc782a. dc782a is actually the ADC EVM that will be used to test the
software setup.

This commit also adds other supporting files and changes like in the makefiles
and removing the header files in the sources.