Modify beaglescope_driver_probe() method
The beaglescope_driver is a rpmsg client driver and registers a iio device on
getting probed. In the earlier versions, he APIs that were used to get/allocate
the iio driver and buffer were not managed APIs.

This commit replaces the non managed APIs with managed APIs (devm_* based APIs).

The driver still do not uses manged API to register the iio_device. This is
because if managed APIs were used and the rpmsg device is put back, it will
also unregister the iio device without taking care if the userspace interface
files exposed iio device are still open, and using these open files will
produce undesirable results.

This commit also places the call to init_waitqueue_head() before the call to
iio_device_register(). This way, the driver will actually be ready to serve as
soon as it registers itself to the iio_subsystem. In the previous version, it
might have produced errors if the user interfaces were used before the
init_waitqueue_head() method returns.

Further, there remain this vulnerability, that if the rpmsg device in released,
the userspace interface is still active and they may try to start the PRUs.
In this case the driver might produce undesirable result and/or kernel OOPs.

Further commits will remove this vulnerability.