Add check for availability of rpmsg device
This commit modifies a few methods in the driver which were using the rpmsg
device associated with the PRUs. This modification is basically to check if the
required rpmsg device is still present or has been released. This check removes
the vulnerability that if the rpmsg device has been released, and the user
interface files are still in use, the iio device will not get released, and
thus userspace interface might still try to use the rpmsg device to start the
PRUs. This could produce kernel OOPs.

Hence using the driver is more safe now.